+++ .eslintrc.js
@@ -11,6 +11,7 @@ module.exports = {
     'airbnb/hooks',
     'plugin:eslint-comments/recommended',
     'plugin:compat/recommended',
+    'plugin:import/typescript',
     'plugin:prettier/recommended',
     'prettier/react',
   ],
@@ -20,15 +21,13 @@ module.exports = {
   plugins: ['filenames', 'simple-import-sort', 'sort-destructure-keys'],
   settings: {
     'import/resolver': {
-      node: {
-        extensions: ['.js', '.ts', '.tsx'],
+      typescript: {
+        directory: ['./blocks/*/tsconfig.json', './packages/*/tsconfig.json', './tsconfig.json'],
       },
-    },
-    'import/parsers': {
-      '@typescript-eslint/parser': ['.ts', '.tsx'],
     },
   },
   rules: {
+    curly: ['error', 'all'],
     'max-len': [
       'error',
       {
@@ -42,6 +41,7 @@ module.exports = {
     'arrow-body-style': ['error', 'as-needed'],
     'prefer-arrow-callback': 'error',
     'no-inline-comments': 'error',
+    'no-implicit-coercion': 'error',
     'filenames/match-regex': ['error', /^\.?[a-z\d]+(\.config|\.test)?$/i, true],
     'filenames/match-exported': 'error',
     quotes: ['error', 'single', { avoidEscape: true, allowTemplateLiterals: false }],