+++ blocks/map/src/index.ts
@@ -1,7 +1,10 @@
 import 'leaflet/dist/leaflet.css';
+import '@wesselkuipers/leaflet.markercluster/dist/MarkerCluster.css';
+import '@wesselkuipers/leaflet.markercluster/dist/MarkerCluster.Default.css';
 import './index.css';
 
 import { attach } from '@appsemble/sdk';
+import { MarkerClusterGroup } from '@wesselkuipers/leaflet.markercluster';
 import { CircleMarker, LocationEvent, Map, TileLayer } from 'leaflet';
 
 import createGetters, { BlockActions, BlockParameters, Events } from './createGetters';
@@ -15,13 +18,29 @@ attach<BlockParameters, BlockActions, Events>(
     const get = createGetters(block.parameters);
     const lat = Number(get.lat(data));
     const lng = Number(get.lng(data));
+    const hasExplicitCenter = Number.isFinite(lat) && Number.isFinite(lng);
     const locationMarker = new CircleMarker(null, {
       color: primaryColor,
     });
+
     const map = new Map(node, {
       attributionControl: false,
       layers: [new TileLayer(tileLayer)],
-    })
+      center: hasExplicitCenter ? [lat, lng] : [0, 0],
+      zoom: 16,
+    });
+
+    // Make sure tile layers are properly loaded.
+    // https://github.com/Leaflet/Leaflet/issues/694
+    setTimeout(() => map.invalidateSize(), 0);
+
+    // Cleanup the map when the block gets removed.
+    utils.addCleanup(() => map.remove());
+
+    map
+      /**
+       * When the user has moved the map, fetch new relevant markers.
+       */
       .on('moveend', () => {
         if (block.events?.emit?.move) {
           events.emit.move({
@@ -32,31 +51,52 @@ attach<BlockParameters, BlockActions, Events>(
           });
         }
       })
+
+      /**
+       *  When a location error occurs because location is disabled, show a message to the user.
+       */
       .once('locationerror', error => {
         // See: https://developer.mozilla.org/en-US/docs/Web/API/PositionError
-        if (error.code && error.code === 1) {
+        if (error?.code === 1) {
           utils.showMessage({
             // XXX Implement i18n.
             body: 'Locatie kon niet worden gevonden. Is de locatievoorziening ingeschakeld?',
           });
-          map.setView([0, 0], 18);
         }
-
         // XXX: Handle TIMEOUT. These are thrown in the .locate() call when `watch` is set to true.
       })
-      .on('locationfound', ({ latlng }: LocationEvent) => {
-        if (Number.isNaN(lat) || Number.isNaN(lng)) {
+
+      /**
+       * Center the map once if location is found.
+       */
+      .once('locationfound', ({ latlng }: LocationEvent) => {
+        if (!hasExplicitCenter) {
           map.setView(latlng, 18);
         }
+      })
+
+      /**
+       * Update the location marker when the user has been updated.
+       */
+      .on('locationfound', ({ latlng }: LocationEvent) => {
         locationMarker.setLatLng(latlng).addTo(map);
       })
+
+      // Start locating the map.
       .locate({ watch: true, timeout: 10e3, maximumAge: 60e3 });
-    if (!Number.isNaN(lat) && !Number.isNaN(lng)) {
-      map.setView([lat, lng], 18);
+
+    let cluster: MarkerClusterGroup;
+
+    if (!block.parameters.disableClustering) {
+      cluster = new MarkerClusterGroup({
+        chunkedLoading: true,
+        maxClusterRadius: block.parameters.maxClusterRadius ?? 80,
+      });
+      map.addLayer(cluster);
     }
 
     events.on.data(d => {
-      loadMarkers(d, fetched, get, data, actions, map);
+      loadMarkers(d, fetched, get, data, actions, cluster || map);
     });
   },
 );