+++ packages/react/jest.config.js
@@ -1,5 +1,8 @@
 module.exports = {
   displayName: '@appsemble/react',
+  moduleNameMapper: {
+    '@appsemble/sdk': '@appsemble/sdk/src',
+  },
   setupFilesAfterEnv: ['<rootDir>/jest.setup.ts'],
   snapshotSerializers: ['enzyme-to-json/serializer'],
 };