+++ packages/studio/components/OrganizationInvite/OrganizationInvite.tsx
@@ -48,7 +48,7 @@ export default function OrganizationInvite(): React.ReactElement {
       }
       setSubmitting(false);
     },
-    [intl, organization.id, push, qs],
+    [intl, organization, push, qs],
   );
 
   const onAcceptClick = React.useCallback(() => sendResponse(true), [sendResponse]);
@@ -60,7 +60,7 @@ export default function OrganizationInvite(): React.ReactElement {
 
     axios
       .get(`/api/invites/${token}`)
-      .then(({ data }) => setOrganization(data))
+      .then(({ data }) => setOrganization(data.organization))
       .catch(() => {
         push({ body: intl.formatMessage(messages.invalidInvite), timeout: 0, dismissable: true });
       })