+++ packages/types/index.d.ts
@@ -1,3 +1,21 @@
+import {
+  Action,
+  BaseAction,
+  Block,
+  BodyHTTPMethods,
+  HTTPMethods,
+  LinkAction,
+  LogAction,
+  Message,
+  RequestAction,
+  RequestLikeActionTypes,
+  ResourceCreateAction,
+  ResourceDeleteAction,
+  ResourceGetAction,
+  ResourceQueryAction,
+  ResourceUpdateAction,
+  Theme,
+} from '@appsemble/sdk';
 import { IconName } from '@fortawesome/fontawesome-common-types';
 import { OpenAPIV3 } from 'openapi-types';
 
@@ -50,186 +68,7 @@ export interface Security {
   };
 }
 
-/*
- * HTTP methods that support a request body.
- */
-export type BodyHTTPMethodsUpper = 'PATCH' | 'POST' | 'PUT';
-
-/**
- * HTTP methods that support a request body, but lower case.
- */
-export type BodyHTTPMethodsLower = 'patch' | 'post' | 'put';
-
-/**
- * HTTP methods that support a request body, but all upper case or all lower case..
- */
-export type BodyHTTPMethods = BodyHTTPMethodsUpper | BodyHTTPMethodsLower;
-
-/**
- * Common HTTP methods.
- */
-export type HTTPMethodsUpper = 'DELETE' | 'GET' | BodyHTTPMethodsUpper;
-
-/**
- * Common HTTP methods, but lower case.
- */
-export type HTTPMethodsLower = 'delete' | 'get' | BodyHTTPMethodsLower;
-
-/**
- * Common HTTP methods, but either all upper case or all lower case.
- */
-export type HTTPMethods = HTTPMethodsUpper | HTTPMethodsLower;
-
-/**
- * A color know to Bulma.
- */
-export type BulmaColor =
-  | 'dark'
-  | 'primary'
-  | 'link'
-  | 'info'
-  | 'success'
-  | 'warning'
-  | 'danger'
-  | 'white';
-
-export interface Theme {
-  /**
-   * The color primarily featured in the color scheme.
-   */
-  primaryColor: string;
-
-  /**
-   * The color used for links.
-   */
-  linkColor: string;
-
-  /**
-   * The color used to feature succesful or positive actions.
-   */
-  successColor: string;
-
-  /**
-   * The color used to indicate information.
-   */
-  infoColor: string;
-
-  /**
-   * The color used for elements that require extra attention.
-   */
-  warningColor: string;
-
-  /**
-   * The color used for elements that demand caution for destructive actions.
-   */
-  dangerColor: string;
-
-  /**
-   * The color used in the foreground of the splash screen.
-   */
-  themeColor: string;
-
-  /**
-   * The color used in the background of the splash screen.
-   */
-  splashColor: string;
-
-  /**
-   * The link to the tile layer used for Leaflet maps.
-   */
-  tileLayer: string;
-}
-
-export interface Message {
-  /**
-   * The content of the message to display.
-   */
-  body: string;
-
-  /**
-   * The color to use for the message.
-   */
-  color?: BulmaColor;
-
-  /**
-   * The timeout period for this message (in milliseconds).
-   */
-  timeout?: number;
-
-  /**
-   * Whether or not to show the dismiss button.
-   */
-  dismissable?: boolean;
-}
-
 export type Navigation = 'bottom' | 'left-menu' | 'hidden';
-
-export interface EventParams {
-  emit?: string;
-  listen?: string;
-}
-
-/**
- * A block that is displayed on a page.
- */
-export interface Block<P = any, A = {}, E extends EventParams = Required<EventParams>> {
-  /**
-   * The type of the block.
-   *
-   * A block type follow the format `@organization/name`.
-   * If the organization is _appsemble_, it may be omitted.
-   *
-   * Pattern:
-   * ^(@[a-z]([a-z\d-]{0,30}[a-z\d])?\/)?[a-z]([a-z\d-]{0,30}[a-z\d])$
-   *
-   * Examples:
-   * - `form`
-   * - `@amsterdam/splash`
-   */
-  type: string;
-
-  /**
-   * A [semver](https://semver.org) representation of the block version.
-   *
-   * Pattern:
-   * ^\d+\.\d+\.\d+$
-   */
-  version: string;
-
-  /**
-   * The theme of the block.
-   */
-  theme?: Theme;
-
-  /**
-   * A free form mapping of named paramters.
-   *
-   * The exact meaning of the parameters depends on the block type.
-   */
-  parameters?: P;
-
-  /**
-   * A mapping of actions that can be fired by the block to action handlers.
-   *
-   * The exact meaning of the parameters depends on the block type.
-   */
-  actions?: A;
-
-  /**
-   * Mapping of the events the block can listen to and emit.
-   *
-   * The exact meaning of the parameters depends on the block type.
-   */
-  events?: {
-    listen: Record<E['listen'], string>;
-    emit: Record<E['emit'], string>;
-  };
-
-  /**
-   * A list of roles that are allowed to view this block.
-   */
-  roles?: string[];
-}
 
 /**
  * A collection of hooks that are triggered upon calling a resource actions.
@@ -261,6 +100,11 @@ export interface ResourceCall {
    * The associated hooks with the resource action.
    */
   hooks: ResourceHooks;
+
+  /**
+   * Query parameters to pass along with the request.
+   */
+  query: Record<string, string>;
 }
 
 export interface Resource {
@@ -313,81 +157,6 @@ export interface Resource {
    */
   url?: string;
 }
-
-export interface BaseAction<T extends string> {
-  /**
-   * The type of the action.
-   */
-  type: T;
-
-  /**
-   * A function which can be called to dispatch the action.
-   */
-  dispatch: (data?: any) => Promise<any>;
-}
-
-type RequestLikeActionTypes =
-  | 'request'
-  | 'resource.create'
-  | 'resource.delete'
-  | 'resource.get'
-  | 'resource.query'
-  | 'resource.update';
-
-export interface LinkAction extends BaseAction<'link'> {
-  /**
-   * Get the link that the action would link to if the given data was passed.
-   */
-  href: (data?: any) => string;
-}
-
-export interface LogAction extends BaseAction<'log'> {
-  /**
-   * The logging level.
-   */
-  level: 'info' | 'warn' | 'error';
-}
-
-export interface RequestLikeAction<T extends RequestLikeActionTypes> extends BaseAction<T> {
-  /**
-   * The HTTP method used to make the request.
-   */
-  method: HTTPMethods;
-  /**
-   * The URL to which the request will be made.
-   */
-  url: string;
-}
-
-export type RequestAction = RequestLikeAction<'request'>;
-export type ResourceCreateAction = RequestLikeAction<'resource.create'>;
-export type ResourceDeleteAction = RequestLikeAction<'resource.delete'>;
-export type ResourceGetAction = RequestLikeAction<'resource.get'>;
-export type ResourceQueryAction = RequestLikeAction<'resource.query'>;
-export type ResourceUpdateAction = RequestLikeAction<'resource.update'>;
-
-/**
- * An action that can be called from within a block.
- */
-export type Action =
-  | BaseAction<'dialog'>
-  | BaseAction<'dialog.error'>
-  | BaseAction<'dialog.ok'>
-  | BaseAction<'flow.back'>
-  | BaseAction<'flow.cancel'>
-  | BaseAction<'flow.finish'>
-  | BaseAction<'flow.next'>
-  | BaseAction<'noop'>
-  | BaseAction<'event'>
-  | LinkAction
-  | LogAction
-  | RequestAction
-  | ResourceGetAction
-  | ResourceQueryAction
-  | ResourceCreateAction
-  | ResourceUpdateAction
-  | ResourceDeleteAction
-  | BaseAction<'resource.subscribe'>;
 
 export interface BlobUploadType {
   type?: 'upload';